+++ hapi/src/api/poi-api.js
@@ -1,7 +1,9 @@
 /* eslint-disable func-names */
 import Boom from "@hapi/boom";
+
+import axios from "axios";
 import { db } from "../models/db.js";
-import { IdSpec, PoiSpec, PoiSpecPlus, PoiArraySpec } from "../models/joi-schemas.js";
+import { IdSpec, PoiSpec, PoiSpecPlus, PoiArraySpec,climateReadingSpec } from "../models/joi-schemas.js";
 import {validationError} from "./logger.js";
 
 export const poiApi = {
@@ -105,5 +107,46 @@ export const poiApi = {
       response: { schema: PoiSpecPlus, failAction: validationError },
       notes: "Find a POI Based on the input ID",
     },
+
+    getWeather:{
+        auth: {
+            strategy: "jwt",
+        },
+        handler: async function (request, h) {
+            try {
+                const poi = await db.poiStore.getPOIById(request.params.id);
+                // get the poi
+                if (!poi) {
+                    return Boom.notFound("No poi with this id");
+                }
+                // read the lat long
+                const apiUrl = `https://api.openweathermap.org/data/2.5/weather?lat=${poi.latitude}&lon=${poi.longitude}&appid=${process.env.openWeatherKey}&units=metric`;
+                const result = await axios.get(apiUrl);
+
+                if (result.status !== 200){
+                    return Boom.notFound("No Response from OpenWeather");
+                }
+
+                const reading = result.data;
+
+                const climateReading = {
+                    code: Number(reading.cod),
+                    temperature: Number(reading.main.temp),
+                    pressure: Number(reading.main.pressure),// Number converts the string to a number
+                    windSpeed: Number(reading.wind.speed),// Number converts the string to a number
+                    windDirection: Number(reading.wind.deg),// Number converts the string to a number
+                };
+
+                return climateReading;
+            } catch (err) {
+                return Boom.serverUnavailable("No poi with this id");
+            }
+        },
+        tags: ["api"],
+        description: "Return the Weather conditions at a given POI ",
+        validate: { params: { id: IdSpec }, failAction: validationError },
+       response: { schema: climateReadingSpec, failAction: validationError },
+        notes: "Uses Open Weather API",
+    }
 
 }