+++ svelte/src/lib/CategoryMap.svelte
@@ -2,6 +2,7 @@
 	import 'leaflet/dist/leaflet.css';
 	import { LeafletMap } from '../services/leaflet-map';
 	import { onMount } from 'svelte';
+	// @ts-ignore
 	import { latestPOI, user } from '../stores';
 	import { geotagService } from '../services/geotag-service';
     let userPoi = [];
@@ -30,12 +31,15 @@
 		map.showZoomControl();
         
 		// for each category add the points of interest
+        // @ts-ignore
         userCategories.forEach(async (category)=>{
            // add the layer
+           // @ts-ignore
            map.addLayerGroup(category.title); 
            // add the pois
            userPoi = await geotagService.getPoiList(category._id);     
            // for each point add map location
+           // @ts-ignore
            userPoi.forEach((poi) =>{
             addPoiMarker(map, poi,category.title);
            });
@@ -43,6 +47,7 @@
         map.showLayerControl();
 	});
 
+	// @ts-ignore
 	function addPoiMarker(map, in_Poi,in_Category) {
 		const poiStr = `${in_Poi.name}`;
 		map.addMarker({ lat: in_Poi.latitude, lng: in_Poi.longitude }, poiStr, in_Category);